+++ spec/decoding/result_spec.rb
@@ -41,6 +41,14 @@ module Decoding
       expect(Result.err(123).unwrap(0)).to be(0)
     end
 
+    it "unwrap_err unwraps err values by returning the inner value" do
+      expect(Result.err(123).unwrap_err(0)).to be(123)
+    end
+
+    it "unwrap_err unwraps ok values by returning the default value" do
+      expect(Result.ok(123).unwrap_err(0)).to be(0)
+    end
+
     it "creates new ok value by mapping with a block" do
       expect(Result.ok(123).map { _1 * 2 }).to eql(Result.ok(246))
     end